Create card view for objects
Moved search to object menu, created cards for objects and changed from list view to ScrollView to show cards instead of simple rows for objects

+++ CityGuide/Model/LandmarkObject.swift
@@ -13,7 +13,7 @@ import CoreLocation
 struct LandmarkObject: Hashable, Codable, Identifiable {
     var id: Int
     var name: String
-    var rating: Int
+    var rating: Double
     var description: String
     var isFavorite: Bool
     var isFeatured: Bool
